refactor(server): split registry entry model
