+++ nodelite-server/src/state/registry.rs
@@ -8,22 +8,25 @@ use std::sync::{Arc, RwLock, RwLockReadGuard, RwLockWriteGuard};
 use std::time::Duration;
 
 use chrono::{DateTime, Utc};
-#[cfg(test)]
-use nodelite_proto::DiskUsage;
 use nodelite_proto::{
-    AlertRuleConfig, GeoIpLocation, InspectionConfig, MetricsConfig, NodeIdentity,
-    NodeListIdentity, NodeListItem, NodeListItemView, NodeListSnapshot, NodeSnapshot, NodeStatus,
-    OverviewData,
+    AlertRuleConfig, GeoIpLocation, InspectionConfig, MetricsConfig, NodeIdentity, NodeListItem,
+    NodeListItemView, NodeSnapshot, NodeStatus, OverviewData,
 };
 
-use super::overview::{OverviewNode, build_overview_from_iter};
+use super::overview::build_overview_from_iter;
 use super::session_control::SessionControlHandle;
 use crate::ServerReadiness;
 use crate::alerts::{
-    AlertStatusView, EvaluatedRule, InspectionReport,
-    build_inspection_report as build_alert_inspection_report, evaluate_rules,
+    EvaluatedRule, InspectionReport, build_inspection_report as build_alert_inspection_report,
+    evaluate_rules,
 };
-use crate::handlers::metrics_routes::{PrometheusNode, render_prometheus_metrics_from_iter};
+use crate::handlers::metrics_routes::render_prometheus_metrics_from_iter;
+
+mod entry;
+#[cfg(test)]
+mod heap_estimate;
+
+use entry::{NodeEntry, geoip_fields_from_location};
 
 const REGISTRY_SHARD_COUNT: usize = 32;
 
@@ -47,267 +50,6 @@ impl Default for Registry {
 #[derive(Debug, Default)]
 struct RegistryShard {
     nodes: HashMap<String, NodeEntry>,
-}
-
-/// 单节点的运行态条目。外部响应模型只在 API / snapshot 边界按需组装。
-///
-/// 字符串池优化: `geoip_country`, `geoip_city`, `location_override_country`, `location_override_city`
-/// 使用 Arc<str> 存储,在高重复场景(如 1000 节点同城)大幅降低内存占用。
-#[derive(Debug, Clone)]
-struct NodeEntry {
-    identity: NodeIdentity,
-    remote_ip: Option<String>,
-    geoip_country: Option<Arc<str>>,
-    geoip_city: Option<Arc<str>>,
-    geoip_latitude: Option<f64>,
-    geoip_longitude: Option<f64>,
-    location_override_country: Option<Arc<str>>,
-    location_override_city: Option<Arc<str>>,
-    location_override_latitude: Option<f64>,
-    location_override_longitude: Option<f64>,
-    snapshot: Option<NodeSnapshot>,
-    last_seen: Option<DateTime<Utc>>,
-    latency_ms: Option<u64>,
-    online: bool,
-    active_session_id: Option<u64>,
-    control: Option<SessionControlHandle>,
-}
-
-impl NodeEntry {
-    fn new(
-        session_id: u64,
-        identity: NodeIdentity,
-        remote_ip: Option<String>,
-        geoip: Option<GeoIpLocation>,
-        location_override: Option<GeoIpLocation>,
-        now: DateTime<Utc>,
-        string_pool: &crate::string_pool::StringPool,
-    ) -> Self {
-        let (geoip_country, geoip_city, geoip_latitude, geoip_longitude) =
-            geoip_fields_from_location(geoip.as_ref(), string_pool);
-        let (
-            location_override_country,
-            location_override_city,
-            location_override_latitude,
-            location_override_longitude,
-        ) = geoip_fields_from_location(location_override.as_ref(), string_pool);
-        Self {
-            identity,
-            remote_ip,
-            geoip_country,
-            geoip_city,
-            geoip_latitude,
-            geoip_longitude,
-            location_override_country,
-            location_override_city,
-            location_override_latitude,
-            location_override_longitude,
-            snapshot: None,
-            last_seen: Some(now),
-            latency_ms: None,
-            online: true,
-            active_session_id: Some(session_id),
-            control: None,
-        }
-    }
-
-    fn from_restored_status(
-        mut status: NodeStatus,
-        string_pool: &crate::string_pool::StringPool,
-    ) -> Self {
-        status.online = false;
-        Self {
-            identity: status.identity,
-            remote_ip: status.remote_ip,
-            geoip_country: status.geoip_country.as_ref().map(|s| string_pool.intern(s)),
-            geoip_city: status.geoip_city.as_ref().map(|s| string_pool.intern(s)),
-            geoip_latitude: status.geoip_latitude,
-            geoip_longitude: status.geoip_longitude,
-            location_override_country: status
-                .location_override_country
-                .as_ref()
-                .map(|s| string_pool.intern(s)),
-            location_override_city: status
-                .location_override_city
-                .as_ref()
-                .map(|s| string_pool.intern(s)),
-            location_override_latitude: status.location_override_latitude,
-            location_override_longitude: status.location_override_longitude,
-            snapshot: status.snapshot,
-            last_seen: status.last_seen,
-            latency_ms: status.latency_ms,
-            online: false,
-            active_session_id: None,
-            control: None,
-        }
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    fn register_session(
-        &mut self,
-        session_id: u64,
-        identity: NodeIdentity,
-        remote_ip: Option<String>,
-        geoip: Option<GeoIpLocation>,
-        location_override: Option<GeoIpLocation>,
-        now: DateTime<Utc>,
-        string_pool: &crate::string_pool::StringPool,
-    ) {
-        let (geoip_country, geoip_city, geoip_latitude, geoip_longitude) =
-            geoip_fields_from_location(geoip.as_ref(), string_pool);
-        let (
-            location_override_country,
-            location_override_city,
-            location_override_latitude,
-            location_override_longitude,
-        ) = geoip_fields_from_location(location_override.as_ref(), string_pool);
-        self.identity = identity;
-        self.remote_ip = remote_ip;
-        self.geoip_country = geoip_country;
-        self.geoip_city = geoip_city;
-        self.geoip_latitude = geoip_latitude;
-        self.geoip_longitude = geoip_longitude;
-        self.location_override_country = location_override_country;
-        self.location_override_city = location_override_city;
-        self.location_override_latitude = location_override_latitude;
-        self.location_override_longitude = location_override_longitude;
-        self.online = true;
-        self.last_seen = Some(now);
-        self.latency_ms = None;
-        self.active_session_id = Some(session_id);
-        self.control = None;
-    }
-
-    fn to_status(&self) -> NodeStatus {
-        NodeStatus {
-            identity: self.identity.clone(),
-            remote_ip: self.remote_ip.clone(),
-            geoip_country: self.geoip_country.as_ref().map(|s| s.to_string()),
-            geoip_city: self.geoip_city.as_ref().map(|s| s.to_string()),
-            geoip_latitude: self.geoip_latitude,
-            geoip_longitude: self.geoip_longitude,
-            location_override_country: self
-                .location_override_country
-                .as_ref()
-                .map(|s| s.to_string()),
-            location_override_city: self.location_override_city.as_ref().map(|s| s.to_string()),
-            location_override_latitude: self.location_override_latitude,
-            location_override_longitude: self.location_override_longitude,
-            snapshot: self.snapshot.clone(),
-            last_seen: self.last_seen,
-            latency_ms: self.latency_ms,
-            online: self.online,
-        }
-    }
-
-    fn to_summary(&self) -> NodeListItem {
-        NodeListItem {
-            identity: NodeListIdentity::from(&self.identity),
-            geoip_country: self.geoip_country.as_ref().map(|s| s.to_string()),
-            geoip_city: self.geoip_city.as_ref().map(|s| s.to_string()),
-            geoip_latitude: self.geoip_latitude,
-            geoip_longitude: self.geoip_longitude,
-            location_override_country: self
-                .location_override_country
-                .as_ref()
-                .map(|s| s.to_string()),
-            location_override_city: self.location_override_city.as_ref().map(|s| s.to_string()),
-            location_override_latitude: self.location_override_latitude,
-            location_override_longitude: self.location_override_longitude,
-            snapshot: self.snapshot.as_ref().map(NodeListSnapshot::from),
-            latency_ms: self.latency_ms,
-            online: self.online,
-        }
-    }
-
-    /// 零拷贝构建视图 (Phase 3.2 优化)。
-    ///
-    /// 与 `to_summary()` 的区别:
-    /// - 直接克隆 `Arc<str>` (只增加引用计数,不复制字符串)
-    /// - 序列化时 serde 直接访问 Arc 内部的 str
-    /// - 避免 ~80 KB 字符串克隆 (1000 节点 × 4 字段 × 20 bytes)
-    fn to_summary_view(&self) -> NodeListItemView {
-        NodeListItemView {
-            identity: NodeListIdentity::from(&self.identity),
-            geoip_country: self.geoip_country.clone(),
-            geoip_city: self.geoip_city.clone(),
-            geoip_latitude: self.geoip_latitude,
-            geoip_longitude: self.geoip_longitude,
-            location_override_country: self.location_override_country.clone(),
-            location_override_city: self.location_override_city.clone(),
-            location_override_latitude: self.location_override_latitude,
-            location_override_longitude: self.location_override_longitude,
-            snapshot: self.snapshot.as_ref().map(NodeListSnapshot::from),
-            latency_ms: self.latency_ms,
-            online: self.online,
-        }
-    }
-
-    fn overview_node(&self) -> OverviewNode<'_> {
-        OverviewNode {
-            online: self.online,
-            latency_ms: self.latency_ms,
-            snapshot: self.snapshot.as_ref(),
-        }
-    }
-
-    fn prometheus_node(&self) -> PrometheusNode<'_> {
-        PrometheusNode {
-            identity: &self.identity,
-            snapshot: self.snapshot.as_ref(),
-            last_seen: self.last_seen,
-            latency_ms: self.latency_ms,
-            online: self.online,
-        }
-    }
-}
-
-impl AlertStatusView for NodeEntry {
-    fn node_id(&self) -> &str {
-        &self.identity.node_id
-    }
-
-    fn node_label(&self) -> &str {
-        &self.identity.node_label
-    }
-
-    fn tags(&self) -> &[String] {
-        &self.identity.tags
-    }
-
-    fn snapshot(&self) -> Option<&NodeSnapshot> {
-        self.snapshot.as_ref()
-    }
-
-    fn last_seen(&self) -> Option<DateTime<Utc>> {
-        self.last_seen
-    }
-
-    fn latency_ms(&self) -> Option<u64> {
-        self.latency_ms
-    }
-
-    fn online(&self) -> bool {
-        self.online
-    }
-}
-
-type GeoIpFields = (Option<Arc<str>>, Option<Arc<str>>, Option<f64>, Option<f64>);
-
-#[allow(clippy::type_complexity)]
-fn geoip_fields_from_location(
-    geoip: Option<&GeoIpLocation>,
-    string_pool: &crate::string_pool::StringPool,
-) -> GeoIpFields {
-    match geoip {
-        Some(location) => (
-            Some(string_pool.intern(&location.country)),
-            location.city.as_ref().map(|city| string_pool.intern(city)),
-            location.latitude,
-            location.longitude,
-        ),
-        None => (None, None, None, None),
-    }
 }
 
 impl Registry {
@@ -710,14 +452,11 @@ impl Registry {
     #[cfg(test)]
     pub(super) fn retained_heap_estimates_for_test(
         status: NodeStatus,
-    ) -> (RetainedHeapEstimate, RetainedHeapEstimate) {
-        let string_pool = crate::string_pool::StringPool::new();
-        let previous_summary = NodeListItem::from(&status);
-        let previous =
-            node_status_heap_estimate(&status) + node_list_item_heap_estimate(&previous_summary);
-        let runtime =
-            node_entry_heap_estimate(&NodeEntry::from_restored_status(status, &string_pool));
-        (runtime, previous)
+    ) -> (
+        heap_estimate::RetainedHeapEstimate,
+        heap_estimate::RetainedHeapEstimate,
+    ) {
+        heap_estimate::retained_heap_estimates_for_status(status)
     }
 }
 
@@ -758,141 +497,4 @@ fn read_lock<T>(lock: &RwLock<T>) -> RwLockReadGuard<'_, T> {
 fn write_lock<T>(lock: &RwLock<T>) -> RwLockWriteGuard<'_, T> {
     lock.write()
         .unwrap_or_else(|poisoned| poisoned.into_inner())
-}
-
-#[cfg(test)]
-#[derive(Debug, Clone, Copy, Default)]
-pub(super) struct RetainedHeapEstimate {
-    pub(super) bytes: usize,
-    pub(super) allocations: usize,
-}
-
-#[cfg(test)]
-impl std::ops::Add for RetainedHeapEstimate {
-    type Output = Self;
-
-    fn add(self, other: Self) -> Self {
-        Self {
-            bytes: self.bytes + other.bytes,
-            allocations: self.allocations + other.allocations,
-        }
-    }
-}
-
-#[cfg(test)]
-fn node_entry_heap_estimate(entry: &NodeEntry) -> RetainedHeapEstimate {
-    node_identity_heap_estimate(&entry.identity)
-        + option_string_heap_estimate(&entry.remote_ip)
-        + option_arc_str_heap_estimate(&entry.geoip_country)
-        + option_arc_str_heap_estimate(&entry.geoip_city)
-        + entry
-            .snapshot
-            .as_ref()
-            .map(node_snapshot_heap_estimate)
-            .unwrap_or_default()
-}
-
-#[cfg(test)]
-fn node_status_heap_estimate(status: &NodeStatus) -> RetainedHeapEstimate {
-    node_identity_heap_estimate(&status.identity)
-        + option_string_heap_estimate(&status.remote_ip)
-        + option_string_heap_estimate(&status.geoip_country)
-        + option_string_heap_estimate(&status.geoip_city)
-        + status
-            .snapshot
-            .as_ref()
-            .map(node_snapshot_heap_estimate)
-            .unwrap_or_default()
-}
-
-#[cfg(test)]
-fn node_list_item_heap_estimate(item: &NodeListItem) -> RetainedHeapEstimate {
-    node_list_identity_heap_estimate(&item.identity)
-        + option_string_heap_estimate(&item.geoip_country)
-        + option_string_heap_estimate(&item.geoip_city)
-}
-
-#[cfg(test)]
-fn node_identity_heap_estimate(identity: &NodeIdentity) -> RetainedHeapEstimate {
-    string_heap_estimate(&identity.node_id)
-        + string_heap_estimate(&identity.node_label)
-        + string_heap_estimate(&identity.hostname)
-        + string_heap_estimate(&identity.os)
-        + option_string_heap_estimate(&identity.kernel_version)
-        + option_string_heap_estimate(&identity.cpu_model)
-        + string_heap_estimate(&identity.agent_version)
-        + string_vec_heap_estimate(&identity.tags)
-}
-
-#[cfg(test)]
-fn node_list_identity_heap_estimate(identity: &NodeListIdentity) -> RetainedHeapEstimate {
-    string_heap_estimate(&identity.node_id)
-        + string_heap_estimate(&identity.node_label)
-        + string_heap_estimate(&identity.hostname)
-        + string_vec_heap_estimate(&identity.tags)
-}
-
-#[cfg(test)]
-fn node_snapshot_heap_estimate(snapshot: &NodeSnapshot) -> RetainedHeapEstimate {
-    vec_buffer_heap_estimate::<DiskUsage>(snapshot.disks.capacity())
-        + snapshot
-            .disks
-            .iter()
-            .map(disk_usage_heap_estimate)
-            .fold(RetainedHeapEstimate::default(), |total, next| total + next)
-}
-
-#[cfg(test)]
-fn disk_usage_heap_estimate(disk: &DiskUsage) -> RetainedHeapEstimate {
-    string_heap_estimate(&disk.device)
-        + string_heap_estimate(&disk.mount_point)
-        + string_heap_estimate(&disk.fs_type)
-}
-
-#[cfg(test)]
-fn string_vec_heap_estimate(values: &[String]) -> RetainedHeapEstimate {
-    vec_buffer_heap_estimate::<String>(values.len())
-        + values
-            .iter()
-            .map(string_heap_estimate)
-            .fold(RetainedHeapEstimate::default(), |total, next| total + next)
-}
-
-#[cfg(test)]
-fn option_string_heap_estimate(value: &Option<String>) -> RetainedHeapEstimate {
-    value.as_ref().map(string_heap_estimate).unwrap_or_default()
-}
-
-#[cfg(test)]
-fn option_arc_str_heap_estimate(value: &Option<Arc<str>>) -> RetainedHeapEstimate {
-    value
-        .as_ref()
-        .map(arc_str_heap_estimate)
-        .unwrap_or_default()
-}
-
-#[cfg(test)]
-fn string_heap_estimate(value: &String) -> RetainedHeapEstimate {
-    RetainedHeapEstimate {
-        bytes: value.capacity(),
-        allocations: usize::from(value.capacity() > 0),
-    }
-}
-
-#[cfg(test)]
-fn arc_str_heap_estimate(value: &Arc<str>) -> RetainedHeapEstimate {
-    // Arc<str> 内存占用 = Arc 控制块 (引用计数等) + 字符串内容
-    // Arc 控制块大约 16-24 字节(取决于平台),字符串内容按实际长度计算
-    RetainedHeapEstimate {
-        bytes: std::mem::size_of::<usize>() * 2 + value.len(),
-        allocations: 1,
-    }
-}
-
-#[cfg(test)]
-fn vec_buffer_heap_estimate<T>(capacity: usize) -> RetainedHeapEstimate {
-    RetainedHeapEstimate {
-        bytes: capacity * std::mem::size_of::<T>(),
-        allocations: usize::from(capacity > 0),
-    }
 }
